requirements.txt
fastapi
uvicorn
jinja2
python-docx
nltk
pyspellchecker
python-multipart